+++ libraries/muxer/src/test/java/androidx/media3/muxer/Mp4MuxerEndToEndTest.java
@@ -15,6 +15,8 @@
  */
 package androidx.media3.muxer;
 
+import static androidx.media3.common.util.Util.getPcmFormat;
+import static androidx.media3.common.util.Util.isPcmEncodingBigEndian;
 import static androidx.media3.muxer.Mp4Muxer.FILE_FORMAT_MP4_WITH_AUXILIARY_TRACKS_EXTENSION;
 import static androidx.media3.muxer.Mp4Muxer.LAST_SAMPLE_DURATION_BEHAVIOR_SET_FROM_END_OF_STREAM_BUFFER_OR_DUPLICATE_PREVIOUS;
 import static androidx.media3.muxer.MuxerTestUtil.FAKE_AUDIO_FORMAT;
@@ -22,6 +24,10 @@ import static androidx.media3.muxer.MuxerTestUtil.FAKE_VIDEO_FORMAT;
 import static androidx.media3.muxer.MuxerTestUtil.XMP_SAMPLE_DATA;
 import static androidx.media3.muxer.MuxerTestUtil.feedInputDataToMuxer;
 import static androidx.media3.muxer.MuxerTestUtil.getFakeSampleAndSampleInfo;
+import static androidx.media3.test.utils.TestUtil.buildDoubleTestSamples;
+import static androidx.media3.test.utils.TestUtil.buildFloatTestSamples;
+import static androidx.media3.test.utils.TestUtil.createByteArray;
+import static androidx.media3.test.utils.TestUtil.createByteBuffer;
 import static com.google.common.base.Preconditions.checkNotNull;
 import static com.google.common.truth.Truth.assertThat;
 import static org.junit.Assert.assertThrows;
@@ -43,21 +49,27 @@ import androidx.media3.extractor.text.DefaultSubtitleParserFactory;
 import androidx.media3.test.utils.DumpFileAsserts;
 import androidx.media3.test.utils.DumpableMp4Box;
 import androidx.media3.test.utils.FakeExtractorOutput;
+import androidx.media3.test.utils.FakeTrackOutput;
 import androidx.media3.test.utils.TestUtil;
 import androidx.test.core.app.ApplicationProvider;
-import androidx.test.ext.junit.runners.AndroidJUnit4;
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.Iterables;
+import com.google.testing.junit.testparameterinjector.TestParameter;
+import com.google.testing.junit.testparameterinjector.TestParameterValuesProvider;
+import java.io.File;
 import java.io.FileOutputStream;
 import java.nio.ByteBuffer;
+import java.nio.ByteOrder;
+import java.util.Random;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
 import org.junit.runner.RunWith;
+import org.robolectric.RobolectricTestParameterInjector;
 import org.robolectric.annotation.Config;
 
 /** End to end tests for {@link Mp4Muxer}. */
-@RunWith(AndroidJUnit4.class)
+@RunWith(RobolectricTestParameterInjector.class)
 public class Mp4MuxerEndToEndTest {
   @Rule public final TemporaryFolder temporaryFolder = new TemporaryFolder();
 
@@ -352,6 +364,34 @@ public class Mp4MuxerEndToEndTest {
     assertThat(extractedFormat.sampleMimeType).isEqualTo(MimeTypes.AUDIO_E_AC3);
     assertThat(extractedFormat.initializationData).hasSize(1);
     assertThat(extractedFormat.initializationData.get(0)).isEqualTo(expectedDec3Payload);
+  }
+
+  @Test
+  public void createMp4File_withAudioPcmFloat_writesAndExtractsCorrectSamples(
+      @TestParameter(valuesProvider = FloatPcmEncodingValuesProvider.class) @C.PcmEncoding
+          int pcmEncoding)
+      throws Exception {
+    File outputFile = temporaryFolder.newFile();
+    Format pcmFloatFormat = getPcmFormat(pcmEncoding, /* channels= */ 1, /* sampleRate= */ 48_000);
+    ByteBuffer sampleBuffer = getTestSamplesForEncoding(pcmEncoding);
+    byte[] sampleData = createByteArray(sampleBuffer.duplicate());
+    BufferInfo bufferInfo =
+        new BufferInfo(/* presentationTimeUs= */ 0L, sampleData.length, C.BUFFER_FLAG_KEY_FRAME);
+
+    try (Mp4Muxer mp4Muxer =
+        new Mp4Muxer.Builder(SeekableMuxerOutput.of(outputFile.getPath())).build()) {
+      int audioTrack = mp4Muxer.addTrack(pcmFloatFormat);
+      mp4Muxer.writeSampleData(audioTrack, sampleBuffer, bufferInfo);
+    }
+
+    FakeExtractorOutput extractorOutput =
+        TestUtil.extractAllSamplesFromFilePath(
+            new Mp4Extractor(new DefaultSubtitleParserFactory()), outputFile.getPath());
+    FakeTrackOutput trackOutput = extractorOutput.trackOutputs.valueAt(0);
+    Format extractedFormat = checkNotNull(trackOutput.lastFormat);
+    assertThat(extractedFormat.sampleMimeType).isEqualTo(MimeTypes.AUDIO_RAW);
+    assertThat(extractedFormat.pcmEncoding).isEqualTo(pcmEncoding);
+    assertThat(trackOutput.getSampleData(/* index= */ 0)).isEqualTo(sampleData);
   }
 
   @Test
@@ -1357,5 +1397,28 @@ public class Mp4MuxerEndToEndTest {
     mp4Muxer.close();
 
     assertThrows(IllegalStateException.class, mp4Muxer::close);
+  }
+
+  private static ByteBuffer getTestSamplesForEncoding(@C.PcmEncoding int pcmEncoding) {
+    Random random = new Random(/* seed= */ 0);
+    ByteBuffer sampleBuffer;
+    if (pcmEncoding == C.ENCODING_PCM_DOUBLE || pcmEncoding == C.ENCODING_PCM_DOUBLE_BIG_ENDIAN) {
+      sampleBuffer = createByteBuffer(buildDoubleTestSamples(/* length= */ 100, random));
+    } else {
+      sampleBuffer = createByteBuffer(buildFloatTestSamples(/* length= */ 100, random));
+    }
+    return sampleBuffer.order(
+        isPcmEncodingBigEndian(pcmEncoding) ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
+  }
+
+  private static final class FloatPcmEncodingValuesProvider extends TestParameterValuesProvider {
+    @Override
+    protected ImmutableList<?> provideValues(TestParameterValuesProvider.Context context) {
+      return ImmutableList.of(
+          value(C.ENCODING_PCM_FLOAT).withName("ENCODING_PCM_FLOAT"),
+          value(C.ENCODING_PCM_FLOAT_BIG_ENDIAN).withName("ENCODING_PCM_FLOAT_BIG_ENDIAN"),
+          value(C.ENCODING_PCM_DOUBLE).withName("ENCODING_PCM_DOUBLE"),
+          value(C.ENCODING_PCM_DOUBLE_BIG_ENDIAN).withName("ENCODING_PCM_DOUBLE_BIG_ENDIAN"));
+    }
   }
 }